Initial import of the kdevelop-pg Java parser.
The parser is feature complete and accepts Java 1.5.
The lack of backtracking or LL(k) lookahead in kdevelop-pg requires a less
elegant solution in form of the java_lookahead class, for the time being,
which works well nevertheless.
The lexer is handicapped because (f)lex can't handle unicode.

There is still a bug in kdevelop-pg which causes trouble
(the symbol handling in reduce_to_epsilon() producing false positives),
which for example makes the parser reject empty parameter lists.
When this is fixed, parsing should work just fine.

CCMAIL: roberto@kdevelop.org