A full pre-processor for the C# parser.
There is now a seperate pre-processor parser (csharp_pp) that works together
with the lexer so that the main parser only gets to see real tokens, with
no disruption from the pre-processor.

This commit also throws out the parser class's i/o actions from main.cpp
into a seperate, to-be-adapted-to-the-framework file called csharp-io.cpp.

So much for the foundations, now I can start with the actual C# grammar...