A rather hackish solution to solve the difficult conflict that made expressions like "a is SomeType ? if_exp : else_exp" impossible. (Nullable expressions stole the question mark from the conditional expression.)
As a result, 99.94% of my test case collection pass.
I believe that 6 of the remaining 11 files are invalid (research pending).
3 others are using an unspecified extension (which is supported by
MS and Mono though, so I'm going to implement it sooner or later).
One file shows that UTF-8 recognition is still not perfect,
and the last one is a virtually unsolvable very-last-corner case.

What I'm saying is that for now, I'm satisfied with the
parser's correctness, and now I'm going for real KDevelop code.
(C# part, here we come!)