Switch kdevelop-pg from Autotools to CMake.
Generating from the example grammars doesn't work with
srcdir != builddir due to a bug in CMake 2.4.2.
I'll be testing with 2.4.3 shortly.
Till then, please build kdev-pg with srcdir == builddir,
as it was before with Autotools.

Don't forget to wipe out all the Autotools files,
in order to feel much better afterwards.