Update the pre-generated files so bison/flex are optional again
Seems like the last changes to the parser-files for kdev-pg's own language
haven't made it into the pre-generated files. Hence it was necessary to
have bison+flex installed to get the -parser/-lexer source code that the
rest of the sources expect. I've copied those generated files from the
builddir back into the source, so this requirement is lifted again.

CCMAIL: git@the-user.org
CCMAIL: kdevelop-devel@kdevelop.org