KDevPG::GNFA::range(begin, end): add missing break after TUcs2 handling
Support for TUcs2 was added in d35cdf7f5aff1910f32cc614de9e89b07080bacd,
with an following empty statement but with break for TUtf16, so effectively
worked.  When later support for TUtf16 was implemented in
20ae735ac64f05057618dcfb2c9b8a3092d1938d, the missing break was missed out.

Seems not have been hit in real world usage so far, but the compiler now
hinted to the implicit fallthrough.